requirements.txt
# ATC Marketing Asset Generator v1.0
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# AI/ML
google-generativeai==0.8.3
langgraph==0.2.60
langchain-google-genai==2.0.7

# Google APIs
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Frontend
streamlit==1.41.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1
pillow==11.0.0
aiofiles==24.1.0

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
